Base class for mypy
